# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GameRAGKit API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/ask/stream": {
      "post": {
        "summary": "Ask an NPC with SSE streaming",
        "operationId": "askStream",
        "parameters": [
          {
            "name": "X-GameRAG-Protocol",
            "in": "header",
            "required": true,
            "schema": { "type": "string", "enum": ["1"] }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "X-Api-Key",
            "in": "header",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AskRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Server Sent Events stream",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "description": "Each line is emitted as `data: <token>\\n\\n`."
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AskRequest": {
        "type": "object",
        "required": ["npc", "question"],
        "properties": {
          "npc": { "type": "string" },
          "question": { "type": "string" },
          "options": { "$ref": "#/components/schemas/AskOptions" }
        }
      },
      "AskOptions": {
        "type": "object",
        "properties": {
          "topK": { "type": "integer", "minimum": 1, "default": 4 },
          "inCharacter": { "type": "boolean", "default": true },
          "systemOverride": { "type": "string" },
          "importance": { "type": "number", "format": "double", "default": 0.2 },
          "forceLocal": { "type": "boolean" },
          "forceCloud": { "type": "boolean" }
        }
      },
...
    }
  }
}